+++ src/lib/Checkbox.svelte
@@ -27,15 +27,20 @@
 		display: flex;
 		align-items: center;
 		justify-content: center;
-		/* TODO wait should this JUST set variables, and buttonlike uses them? yes! */
 		border-top: var(--border_width_lg) var(--border_style) var(--border_lighter_color);
 		border-bottom: var(--border_width_lg) var(--border_style) var(--border_lighter_color);
+		--outline_color: transparent;
 	}
 	.selected {
+		/* TODO vars */
 		border-color: var(--selected_color);
 	}
+	.selected:focus {
+		--outline_color: var(--selected_color);
+	}
 	.selected:active {
-		border-color: var(--border_active_color);
+		border-color: var(--active_color);
+		--outline_color: var(--active_color);
 	}
 	.content {
 		width: var(--column_width);