+++ src/onboard/Credits.svelte
@@ -1,6 +1,6 @@
 <strong>TODO proper credits</strong>
 <p>
-	this study is inspired by <a href="https://consentful.systems">consentful.systems</a>
+	this sketch is inspired by <a href="https://consentful.systems">consentful.systems</a>
 </p>
 <p>
 	permissive open source code at <a href="https://github.com/feltcoop/felt">github</a> 💚