+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## 0.1.5
 
-- add first pass at onboarding study
+- add first pass at onboarding sketch
   ([#32](https://github.com/feltcoop/felt/pull/32))
 
 ## 0.1.4